fix: 기존 Cleric 으로 수정
- 새로 만든 폴더 삭제
- 기존 클레릭에 생성자 추가
